+++ natural-language-understanding/src/main/java/com/ibm/watson/developer_cloud/natural_language_understanding/v1/model/RelationsResult.java
@@ -24,7 +24,7 @@ import com.ibm.watson.developer_cloud.service.model.GenericModel;
 public class RelationsResult extends GenericModel {
 
   /** Confidence score for the relation. Higher values indicate greater confidence. */
-  private Long score;
+  private Double score;
   /** The sentence that contains the relation. */
   private String sentence;
   /** The type of the relation. */
@@ -40,7 +40,7 @@ public class RelationsResult extends GenericModel {
    * @param type The type of the relation.
    * @param arguments The extracted relation objects from the text.
    */
-  public RelationsResult(final Long score, final String sentence, final String type, final List<RelationArgument> arguments) {
+  public RelationsResult(final Double score, final String sentence, final String type, final List<RelationArgument> arguments) {
     this.score = score;
     this.sentence = sentence;
     this.type = type;
@@ -61,7 +61,7 @@ public class RelationsResult extends GenericModel {
    *
    * @return the score
    */
-  public Long getScore() {
+  public Double getScore() {
     return score;
   }
 
@@ -97,7 +97,7 @@ public class RelationsResult extends GenericModel {
    *
    * @param score the new score
    */
-  public void setScore(final Long score) {
+  public void setScore(final Double score) {
     this.score = score;
   }
